se agrego eliminacion de la imagenes del producto

+++ back/services/products-service.js
@@ -134,6 +134,20 @@ async function getProductosPorVendedor(id) {
 // eliminar articulo
 async function deleteArticle(id) {
   try {
+    const productoConImagen = await Products.findByPk(id, {
+      include: [{ model: Images }],
+    });
+
+    if (!productoConImagen) {
+      return { success: false, message: `No se encontró el Articulo con ID ${id}.` };
+    }
+
+    if (productoConImagen.Images) {
+      await Promise.all(productoConImagen.Images.map(async (imagen) => {
+        await imagen.destroy();
+      }));
+    }
+
     const result = await Products.destroy({
       where: {
         id,
@@ -143,6 +157,7 @@ async function deleteArticle(id) {
     if (result === 1) {
       return { success: true, message: `Articulo con ID ${id} eliminado exitosamente.` };
     }
+
     return { success: false, message: `No se encontró el Articulo con ID ${id}.` };
   } catch (error) {
     console.error(`Error al eliminar el articulo: ${error.message}`);
